fix: move options
