+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,99 +1,6 @@
-import jwt_decode from "jwt-decode";
-import NextAuth, { Account, NextAuthOptions, Session } from "next-auth";
-import { JWT } from "next-auth/jwt";
-import KeycloakProvider from "next-auth/providers/keycloak";
+import NextAuth from "next-auth";
 
-import { encrypt } from "@/components/auth/encryption";
-
-// this will refresh an expired access token, when needed
-
-async function refreshAccessToken(token: { refresh_token: string }) {
-  const resp = await fetch(`${process.env.REFRESH_TOKEN_URL}`, {
-    headers: { "Content-Type": "application/x-www-form-urlencoded" },
-    body: new URLSearchParams({
-      client_id: process.env.DEMO_FRONTEND_CLIENT_ID!,
-      client_secret: process.env.DEMO_FRONTEND_CLIENT_SECRET!,
-      grant_type: "refresh_token",
-      refresh_token: token.refresh_token,
-    }),
-    method: "POST",
-  });
-  const refreshToken = (await resp.json()) as {
-    access_token: string;
-    id_token: string;
-    expires_in: number;
-    refresh_token: string;
-  };
-  if (!resp.ok) throw refreshToken;
-
-  return {
-    ...token,
-    access_token: refreshToken.access_token,
-    decoded: jwt_decode(refreshToken.access_token),
-    id_token: refreshToken.id_token,
-    expires_at: Math.floor(Date.now() / 1000) + refreshToken.expires_in,
-    refresh_token: refreshToken.refresh_token,
-  };
-}
-
-const authOptions: NextAuthOptions = {
-  providers: [
-    KeycloakProvider({
-      clientId: `${process.env.DEMO_FRONTEND_CLIENT_ID}`,
-      clientSecret: `${process.env.DEMO_FRONTEND_CLIENT_SECRET}`,
-      issuer: `${process.env.AUTH_ISSUER}`,
-    }),
-  ],
-
-  callbacks: {
-    async jwt({ token, account }: { token: JWT; account: Account | null }) {
-      const nowTimeStamp = Math.floor(Date.now() / 1000);
-
-      if (account?.access_token) {
-        // account is only available the first time this callback is called on a new session (after the user signs in)
-
-        token.decoded = jwt_decode(account.access_token);
-        token.access_token = account.access_token;
-        token.id_token = account.id_token;
-        token.expires_at = account.expires_at;
-        token.refresh_token = account.refresh_token;
-        return token;
-      } else if (nowTimeStamp < Number(token.expires_at)) {
-        // token has not expired yet, return it
-        return token;
-      } else {
-        // token is expired, try to refresh it
-        console.log("Token has expired. Will refresh...");
-        try {
-          const refreshedToken = await refreshAccessToken(
-            token as { refresh_token: string },
-          );
-          console.log("Token is refreshed.");
-          return refreshedToken;
-        } catch (error) {
-          console.error("Error refreshing access token", error);
-          return { ...token, error: "RefreshAccessTokenError" };
-        }
-      }
-    },
-    session({ session: session_, token }: { session: Session; token: JWT }) {
-      // Send properties to the client
-      const session = session_ as Session & {
-        access_token: string;
-        id_token: string;
-        roles: string[];
-        error: unknown;
-      };
-      session.access_token = encrypt(token.access_token as string); // see utils/sessionTokenAccessor.js
-      session.id_token = encrypt(token.id_token as string); // see utils/sessionTokenAccessor.js
-      session.roles = (
-        token.decoded as { realm_access: { roles: string[] } }
-      ).realm_access.roles;
-      session.error = token.error;
-      return session;
-    },
-  },
-};
+import { authOptions } from "@/app/api/auth/options";
 
 const handler = NextAuth(authOptions) as typeof NextAuth;
