+++ src/app/api/auth/logout/route.ts
@@ -1,7 +1,7 @@
 import { getServerSession } from "next-auth";
 
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 import { getIdToken } from "@/components/auth/sessionTokenAccessor";
+import { authOptions } from "@/app/api/auth/options";
 
 export async function GET() {
   const session = await getServerSession(authOptions);
